# NewAPI/bin/Release/net7.0/swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "newapi",
    "version": "v1"
  },
  "paths": {
    "/api/AwsS3": {
      "get": {
        "tags": [
          "AwsS3"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "documentName",
            "type": "string",
            "collectionFormat": "multi"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      },
      "post": {
        "tags": [
          "AwsS3"
        ],
        "consumes": [
          "multipart/form-data"
        ],
        "parameters": [
          {
            "in": "formData",
            "name": "file",
            "type": "file"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      },
      "delete": {
        "tags": [
          "AwsS3"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "documentName",
            "type": "string",
            "collectionFormat": "multi"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
...
  },
...
  "securityDefinitions": {
    "Bearer": {
      "type": "apiKey",
      "name": "Authorization",
      "in": "header",
      "description": "Header de autorização JWT usando o esquema Bearer.\r\n\r\nInforme 'Bearer'[espaço] e o seu token.\r\n\r\nExamplo: 'Bearer 12345abcdef'"
    }
  },
  "security": [
    {
      "Bearer": [ ]
    }
  ]
}